third_party/WebKit/LayoutTests/fast/css/invalidation/read-only-write-pseudo-expected.txt

Use descendant invalidation sets for :read-only and :read-write pseudo classes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(i1, '').backgroundColor is transparent
PASS getComputedStyle(i2, '').backgroundColor is transparent
FAIL internals.updateStyleAndReturnAffectedElementCount() should be 1. Was 2.
PASS getComputedStyle(i1, '').backgroundColor is green
FAIL internals.updateStyleAndReturnAffectedElementCount() should be 1. Was 2.
PASS getComputedStyle(i2, '').backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
